% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLFleet-method.R
\docType{methods}
\name{as.data.frame,DEPLFleet-method}
\alias{as.data.frame,DEPLFleet-method}
\title{as.data.frame}
\usage{
\S4method{as.data.frame}{DEPLFleet}(x, row.names = NULL, optional = FALSE,
  ...)
}
\description{
Create a \code{data.frame} from the \code{DEPLFleet} object
}

