% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLPar-method.R
\docType{methods}
\name{as.data.frame,DEPLPar-method}
\alias{as.data.frame,DEPLPar-method}
\title{as.data.frame}
\usage{
\S4method{as.data.frame}{DEPLPar}(x, row.names = NULL, optional = FALSE,
  ...)
}
\description{
Create a \code{data.frame} from the \code{DEPLPar} object
}

