% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLSim-method.R
\docType{methods}
\name{catch.n,DEPLSim-method}
\alias{catch.n,DEPLSim-method}
\title{catch.n}
\usage{
\S4method{catch.n}{DEPLSim}(object)
}
\value{
Array
}
\description{
Accessor of the \code{catch.n} slot of a \code{DEPLSim} object
}
\section{Slots}{

\describe{
\item{\code{object}}{of \code{DEPLSim} class}
}}

