% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLFleet-method.R
\docType{methods}
\name{catchability<-,DEPLFleet-method}
\alias{catchability<-,DEPLFleet-method}
\title{catchability<-}
\usage{
\S4method{catchability}{DEPLFleet}(x) <- value
}
\description{
Accessor of the catchability of a \code{DEPLFleet} object
}

