% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLFleet-class.R
\docType{class}
\name{DEPLFleet-class}
\alias{DEPLFleet-class}
\title{DEPLFleet class and methods}
\description{
A class for creating an object that describes a fleet
}
\section{Slots}{

\describe{
\item{\code{n}}{\code{integer}. The number of fleets.}

\item{\code{name}}{\code{character} or \code{vector} of characters describing the fleets names}

\item{\code{catchability}}{\code{numeric} or vector giving the catchability of each fleet}
}}
\section{Constructor}{
 
 The DEPLFleet(object) constructor method allows simple creation of new DEPLFleet object.
}
\examples{
DEPLFleet(n=1,name="artisanal",catchability=3.2E-5)
}
\seealso{
DEPLSim, DEPLPar, DEPLZone.
}

