% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLFleet.R
\name{DEPLFleet}
\alias{DEPLFleet}
\title{DEPLFleet constructor}
\usage{
DEPLFleet(n, name, catchability)
}
\description{
Creates an object of the DEPLFleet class representing a fleet exploiting the fishery.
}
\section{Slots}{

\describe{
\item{\code{n}}{\code{integer}. The number of fleets.}

\item{\code{name}}{\code{character} or \code{vector} of characters describing the fleets names}

\item{\code{catchability}}{\code{numeric} or vector giving the catchability of each fleet}
}}
\examples{
DEPLFleet(n=1,name="artisanal",catchability=3.2E-5)
}

