% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLPar-class.R
\docType{class}
\name{DEPLPar-class}
\alias{DEPLPar-class}
\title{DEPLPar class and methods}
\value{
\code{DEPLPar} object
}
\description{
A class for storing the parameters of a DEPLSim model
}
\section{Slots}{

\describe{
\item{\code{value}}{\code{numeric} or \code{vector} giving the values of parameters}

\item{\code{name}}{\code{character} or \code{vector} of characters describing the parameter names}

\item{\code{unit}}{\code{character} or \code{vector} of characters describing the units of parameters}
}}
\section{Constructor}{

 The DEPLPar(object) constructor method allows simple creation of new DEPLPar object.
}
\examples{
DEPLPar(value=3.0E4,name="Biomass",unit="t")
}
\seealso{
DEPLZone, DEPLFleet, DEPLSim
}

