% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLPar.R
\name{DEPLPar}
\alias{DEPLPar}
\title{DEPLPar constructor.}
\usage{
DEPLPar(value, name, unit)
}
\description{
Creates an object of the DEPLPar class storing simulation parameters.
}
\section{Slots}{

\describe{
\item{\code{value}}{\code{numeric} or \code{vector} giving the values of parameters}

\item{\code{name}}{\code{character} or \code{vector} of characters describing the parameter names}

\item{\code{unit}}{\code{character} or \code{vector} of characters describing the units of parameters}
}}
\examples{
DEPLPar(value=3.0E4,name="Biomass",unit="t")
}

