% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLSim-class.R
\docType{class}
\name{DEPLSim-class}
\alias{DEPLSim-class}
\title{DEPLSim class and methods}
\value{
\code{DEPLSim} object
}
\description{
A class that represents the Simulation Depeltion model for Octopus.
Creates an object of the DEPLSim class that implements the Simulation Depletion model.
}
\section{Slots}{

\describe{
\item{\code{\code{controle}}}{\code{DEPLPar}}

\item{\code{\code{params}}}{\code{DEPLPar}}

\item{\code{\code{zones}}}{\code{DEPLZone}}

\item{\code{\code{catch.n}}}{\code{array}}

\item{\code{\code{catch.wt}}}{\code{array}}

\item{\code{\code{stock.n}}}{\code{array}}

\item{\code{\code{stock.wt}}}{\code{array}}

\item{\code{\code{cpue}}}{\code{array}}
}}
\author{
Abdelouahed BEN MHAMED
}
\seealso{
DEPLPar, DEPLZone, DEPLFleet
}

