% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLSim.R
\name{DEPLSim}
\alias{DEPLSim}
\title{DEPLSim class and methods}
\usage{
DEPLSim(file = "missing")
}
\description{
A class for creating a simulation object under DEPLr
}
\section{Slots}{

\describe{
\item{\code{controle}}{\code{DPELPar}. Controle parameters of the simualtion.}

\item{\code{zones}}{\code{DEPLZone}. Zones in the simulation.}

\item{\code{catch.n}}{\code{array}. An \code{array} that stores the catch in numbers.}

\item{\code{catch.wt}}{\code{array}. A \code{array} that stores the catch in weight.}

\item{\code{stock.n}}{\code{array}. A \code{array} that stores the abundance of the stock.}

\item{\code{stock.wt}}{\code{array}. A \code{array} that stores the biomass of the stock.}

\item{\code{cpue}}{\code{array}. A \code{array} that stores the cpue data.}
}}
\section{Constructor}{
 
 The DEPLSim(object) constructor method allows simple creation of new DEPLSim object.
}

