% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLZone-class.R
\docType{class}
\name{DEPLZone-class}
\alias{DEPLZone-class}
\title{DEPLZone class and methods}
\value{
\code{DEPLZone} object
}
\description{
A class for creating an object that describes a fishing zone
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character}. The name of the zone.}

\item{\code{fleet}}{\code{DEPLFleet}. An object of class DEPLFleet}

\item{\code{initAbund}}{\code{numeric}. Abundance in the begining of the fishing season}

\item{\code{initLength}}{\code{numeric}. Length in the begining of the fishing season}

\item{\code{propRec}}{\code{numeric}. Proportion of total recruitment recieved by zone.}

\item{\code{data}}{\code{array}. Fishing effort, Weekly natural mortality and recruitment distribution.}
}}
\section{Constructor}{

 The DEPLZone(object) constructor method allows simple creation of new DEPLZone object.
}
\seealso{
DEPLPar, DEPLFleet, DEPLSim
}

