% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLZone.R
\name{DEPLZone}
\alias{DEPLZone}
\title{DEPLZone constructor}
\usage{
DEPLZone(n, name, fleet, initAbund, initLength, propRec, data)
}
\description{
Creates an object of the DEPLZone class representing a zone exploited in the fishery.
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character}. The name of the zone.}

\item{\code{fleet}}{\code{DEPLFleet}. An object of class DEPLFleet}

\item{\code{initAbund}}{\code{numeric}. Abundance in the begining of the fishing season}

\item{\code{initLength}}{\code{numeric}. Length in the begining of the fishing season}

\item{\code{propRec}}{\code{numeric}. Proportion of total recruitment recieved by zone.}

\item{\code{data}}{\code{array}}
}}

