% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLZone-method.R
\docType{methods}
\name{describe,DEPLZone-method}
\alias{describe,DEPLZone-method}
\title{describe}
\usage{
\S4method{describe}{DEPLZone}(object)
}
\description{
return a description of a DEPLZone in a data.frame
}

