% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLZone-method.R
\docType{methods}
\name{fleet,DEPLZone-method}
\alias{fleet,DEPLZone-method}
\title{fleet}
\usage{
\S4method{fleet}{DEPLZone}(object)
}
\value{
DEPLFleet object
}
\description{
Return the fleet exploiting the zones.
}

