% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLSim-method.R
\docType{methods}
\name{index,DEPLSim-method}
\alias{index,DEPLSim-method}
\title{index}
\usage{
\S4method{index}{DEPLSim}(object)
}
\value{
Array
}
\description{
Accessor of the \code{cpue} slot of a \code{DEPLSim} object
}
\section{Slots}{

\describe{
\item{\code{object}}{of \code{DEPLSim} class}
}}

