% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sim.R
\name{read.sim}
\alias{read.sim}
\title{A function that allow reading an input simulation file}
\usage{
read.sim(file)
}
\arguments{
\item{x}{: path of the input file}
}
\description{
Read a simulation input file
}

