% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLPar-method.R
\docType{methods}
\name{show,DEPLPar-method}
\alias{show,DEPLPar-method}
\title{show}
\usage{
\S4method{show}{DEPLPar}(object)
}
\description{
Method for printing a DEPLPar object
}
\section{Slots}{

\describe{
\item{\code{object}}{object of DEPLPar class}
}}

