% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLSim-method.R
\docType{methods}
\name{stock.n,DEPLSim-method}
\alias{stock.n,DEPLSim-method}
\title{stock.n}
\usage{
\S4method{stock.n}{DEPLSim}(object)
}
\value{
Array
}
\description{
Accessor of the \code{stock.n} slot of a \code{DEPLSim} object
}
\section{Slots}{

\describe{
\item{\code{object}}{of \code{DEPLSim} class}
}}

