% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLSim-method.R
\docType{methods}
\name{summary,DEPLSim-method}
\alias{summary,DEPLSim-method}
\title{summary}
\usage{
\S4method{summary}{DEPLSim}(object, stat = "character")
}
\description{
Return a summary \code{data.frame} of the object in parameters
}

