% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLSim-method.R
\docType{methods}
\name{zones,DEPLSim-method}
\alias{zones,DEPLSim-method}
\title{zones}
\usage{
\S4method{zones}{DEPLSim}(object)
}
\description{
Return a DEPLZone object from a DEPLSim object
}

