% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPLSim-method.R
\docType{methods}
\name{controle,DEPLSim-method}
\alias{controle,DEPLSim-method}
\title{controle}
\usage{
\S4method{controle}{DEPLSim}(object)
}
\description{
Return the controle parameter of the DEPLSim object
}

