% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPLFleet-method.R
\docType{methods}
\name{show,DEPLFleet-method}
\alias{show,DEPLFleet-method}
\title{show}
\usage{
\S4method{show}{DEPLFleet}(object)
}
\description{
print a DEPLFleet object
}
\section{Slots}{

\describe{
\item{\code{object}}{of DEPLFleet class}
}}

